requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
qdrant-client==1.7.3
google-generativeai==0.3.2
python-dotenv==1.0.0
